feature/24: local dev runner, edit type

+++ src/main/resources/db/changelog/changelogs/script/01-06-create-message-table.sql
@@ -5,7 +5,7 @@ GO
 
 CREATE TABLE message
 (
-    message_id int8 PRIMARY KEY,
+    message_id INTEGER PRIMARY KEY,
     message    jsonb,
     inquiry_id uuid REFERENCES public.inquiry (inquiry_id)
 )
